Temporary hot fix - don't remove test version "v202512" in production

+++ src/datapump/jobs/geotrellis.py
@@ -309,7 +309,8 @@ class GeotrellisJob(Job):
 
                     versions_to_delete = versions[: -GLOBALS.max_versions]
                     for version in versions_to_delete:
-                        client.delete_version(table.dataset, version)
+                        if version != "v202512":
+                            client.delete_version(table.dataset, version)
 
             if (
                 self.table.analysis == Analysis.viirs
